assigning icons the best way
Through the way stated here https://stackoverflow.com/a/10849817/4373371, thus through assigning the resource name (the last part of R.drawable.img1, that is img1) to the tag element of the imagebutton, we can assign an icon to the menu item.
So it is the tag that must be transmitted to clients, and there we can use e.g. setIconOfCheckedNavMenuItem(...) method for each navigation menu item.